Refactor useMovieRecommendations hook
